+++ src/lib/utils/s21_trig_range_reduction.c
@@ -2,7 +2,7 @@
 
 double s21_trig_range_reduction(double x) {
   if (x > S21_PI || x < -S21_PI) {
-    return (double)s21_fmod(x, S21_PI * 2);
+    return (double)s21_fmod(x, (double)(S21_PI * 2));
   }
 
   return x;
